moved classes from api package to core
