+++ src/main/java/nl/knaw/dans/migration/core/DataverseLoaderImpl.java
@@ -16,7 +16,6 @@
 package nl.knaw.dans.migration.core;
 
 import io.dropwizard.hibernate.UnitOfWork;
-import nl.knaw.dans.migration.api.ActualFile;
 import nl.knaw.dans.migration.db.ActualFileDAO;
 import nl.knaw.dans.lib.dataverse.DataverseClient;
