+++ src/main/java/nl/knaw/dans/migration/core/EasyFile.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.migration.api;
+package nl.knaw.dans.migration.core;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
