+++ src/main/java/nl/knaw/dans/migration/core/EasyFileLoader.java
@@ -15,8 +15,6 @@
  */
 package nl.knaw.dans.migration.core;
 
-import nl.knaw.dans.migration.api.EasyFile;
-import nl.knaw.dans.migration.api.ExpectedFile;
 import nl.knaw.dans.migration.db.EasyFileDAO;
 import nl.knaw.dans.migration.db.ExpectedFileDAO;
 import org.apache.commons.csv.CSVFormat;
