+++ src/main/java/nl/knaw/dans/migration/core/ExpectedFile.java
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.migration.api;
+package nl.knaw.dans.migration.core;
 
-import nl.knaw.dans.migration.core.FileRights;
-import nl.knaw.dans.migration.core.ManifestCsv;
 import org.hsqldb.lib.StringUtil;
 import org.joda.time.DateTime;
