+++ src/main/java/nl/knaw/dans/migration/core/ExpectedFileKey.java
@@ -13,27 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.migration.api;
+package nl.knaw.dans.migration.core;
+
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
 
 import java.io.Serializable;
 import java.util.Objects;
 
-public class ActualFileKey implements Serializable {
+public class ExpectedFileKey implements Serializable {
+
   private String doi;
-  private String actualPath;
-  private int majorVersionNr;
-  private int minorVersionNr;
+  private String expectedPath;
+  private int removedDuplicateFileCount;
 
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
-    ActualFileKey that = (ActualFileKey) o;
-    return majorVersionNr == that.majorVersionNr && minorVersionNr == that.minorVersionNr && Objects.equals(doi, that.doi) && Objects.equals(actualPath, that.actualPath);
+    ExpectedFileKey that = (ExpectedFileKey) o;
+    return removedDuplicateFileCount == that.removedDuplicateFileCount && Objects.equals(doi, that.doi) && Objects.equals(expectedPath, that.expectedPath);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(doi, actualPath, majorVersionNr, minorVersionNr);
+    return Objects.hash(doi, expectedPath, removedDuplicateFileCount);
   }
 }
