+++ src/main/java/nl/knaw/dans/migration/db/ActualFileDAO.java
@@ -16,7 +16,7 @@
 package nl.knaw.dans.migration.db;
 
 import io.dropwizard.hibernate.AbstractDAO;
-import nl.knaw.dans.migration.api.ActualFile;
+import nl.knaw.dans.migration.core.ActualFile;
 import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
